Mejora en funciones de manejo de archivos.

+++ src/IO.php
@@ -402,7 +402,7 @@ class IO
 
 	public static function HasFiles($path, $ext = '')
 	{
-		if ($handle = opendir($path))
+		if ($handle = self::OpenDirNoWarning($path))
 		{
 			while (false !== ($entry = readdir($handle)))
 			{
@@ -512,7 +512,7 @@ class IO
 			}
 			$exclusions = $exclusionsFull;
 		}
-		$dir_handle = opendir($dirsource);
+		$dir_handle = self::OpenDirNoWarning($dirsource);
 
 		while($file = readdir($dir_handle))
 		{
@@ -546,7 +546,7 @@ class IO
 		// recursive function to copy all subdirectories and contents
 		$dir_handle = null;
 		if(is_dir($dirsource))
-			$dir_handle = opendir($dirsource);
+			$dir_handle = self::OpenDirNoWarning($dirsource);
 		if ($dirname == '')
 			$dirname = substr($dirsource, strrpos($dirsource, '/') + 1);
 
@@ -585,6 +585,10 @@ class IO
 		return true;
 	}
 
+	/**
+	 * Remueve directorio completo aunque
+	 * contenga archivos.
+	 */
 	public static function RemoveDirectory($dir)
 	{
 		if (!file_exists($dir))
@@ -595,7 +599,7 @@ class IO
 			return 1;
 		}
 		$n = 0;
-		if($dh = opendir($dir))
+		if($dh = self::OpenDirNoWarning($dir))
 		{
 			while(($file = readdir($dh)) !== false)
 			{
@@ -604,9 +608,44 @@ class IO
 				$n += self::RemoveDirectory($dir.'/'.$file);
 			}
 			closedir($dh);
-			rmdir($dir);
+			self::RmDir($dir);
 		}
 		return $n;
+	}
+
+	/**
+	 * Wrapper de función rmdir de php,
+	 * para evitar warnings.
+	 * Sólo borra directorios vacíos.
+	 */
+	public static function RmDir($dir)
+	{
+		try
+		{
+			if(file_exists($dir))
+				return rmdir($dir);
+		}
+		catch(\Exception $e)
+		{
+			if($e->getCode() !== E_WARNING)
+				Log::HandleSilentException($e);
+		}
+		return false;
+	}
+
+	public static function OpenDirNoWarning($dir)
+	{
+		try
+		{
+			if(file_exists($dir))
+				return opendir($dir);
+		}
+		catch(\Exception $e)
+		{
+			if($e->getCode() !== E_WARNING)
+				Log::HandleSilentException($e);
+		}
+		return false;
 	}
 
 	public static function GetDirectoryINodesCount($dir)
@@ -646,7 +685,7 @@ class IO
 
 	private static function GetDirectorySizeWin($dir)
 	{
-		if(($dh = opendir($dir)) == false)
+		if(($dh = self::OpenDirNoWarning($dir)) == false)
 		{
 			return array('size' => 0, 'inodes' => 0);
 		}
@@ -729,9 +768,17 @@ class IO
 
 	public static function Move($source, $target)
 	{
-		//Backup::AppendDeleted($source);
-		self::Delete($target);
-		rename($source, $target);
+		try
+		{
+			if(file_exists($source))
+				return rename($source, $target);
+		}
+		catch(\Exception $e)
+		{
+			if($e->getCode() !== E_WARNING)
+				Log::HandleSilentException($e);
+		}
+		return false;
 	}
 
 	public static function IsCompressedDirectory($path)
@@ -774,12 +821,13 @@ class IO
 		try
 		{
 			if (file_exists($file))
-				unlink($file);
+				return unlink($file);
 		}
 		catch(\Exception $e)
 		{
 			if($e->getCode() !== E_WARNING)
 				Log::HandleSilentException($e);
 		}
+		return false;
 	}
 }
